pc/requirements.txt
imageio==2.35.1
imageio-ffmpeg==0.5.1
numpy==2.1.1
pillow==10.4.0
psutil==6.0.0
pyglet==2.0.17
pyserial==3.5
PyYAML==6.0.2
setuptools==75.1.0
